+++ fast/stages/1-tenant-factory/variables-fast.tf
@@ -120,13 +120,15 @@ variable "org_policy_tags" {
   })
 }
 
+check "prefix_validator" {
+  assert {
+    condition     = (try(length(var.prefix), 0) < 10) || (try(length(var.prefix), 0) < 12 && var.root_node != null)
+    error_message = "var.prefix must be 9 characters or shorter for organizations, and 11 chars or shorter for tenants."
+  }
+}
+
 variable "prefix" {
   # tfdoc:variable:source 0-bootstrap
   description = "Prefix used for resources that need unique names. Use 9 characters or less."
   type        = string
-
-  validation {
-    condition     = try(length(var.prefix), 0) < 10
-    error_message = "Use a maximum of 9 characters for prefix."
-  }
 }