+++ fast/stages/2-security/variables.tf
@@ -79,16 +79,6 @@ variable "outputs_location" {
   default     = null
 }
 
-variable "prefix" {
-  # tfdoc:variable:source 0-bootstrap
-  description = "Prefix used for resources that need unique names. Use 9 characters or less."
-  type        = string
-  validation {
-    condition     = try(length(var.prefix), 0) < 10
-    error_message = "Use a maximum of 9 characters for prefix."
-  }
-}
-
 variable "vpc_sc" {
   description = "VPC SC configuration."
   type = object({