+++ tests/fixtures/gcs.tf
@@ -17,6 +17,7 @@ module "gcs" {
   project_id = var.project_id
   prefix     = var.prefix
   name       = "my-bucket"
+  location   = "EU"
   iam = {
     "roles/storage.admin" = ["serviceAccount:service-${var.project_number}@gcp-sa-pubsub.iam.gserviceaccount.com"]
   }